@@ -4,4 +4,23 @@ in-place terminal dashboards with a co-existing scrolling log region.
 """
 from .dashboard import TermDash
 from .components import Line, Stat, AggregatedLine
-from .utils import format_bytes
+from .utils import format_bytes, fmt_hms, bytes_to_mib, clip_ellipsis
+
+# New: progress bar, simple board, cmake-like printer
+from .progress import ProgressBar
+from .simpleboard import SimpleBoard
+from .seemake import SeemakePrinter
+
+__all__ = [
+    "TermDash",
+    "Line",
+    "Stat",
+    "AggregatedLine",
+    "ProgressBar",
+    "SimpleBoard",
+    "SeemakePrinter",
+    "format_bytes",
+    "fmt_hms",
+    "bytes_to_mib",
+    "clip_ellipsis",
+]
