@@ -1,561 +1,1048 @@
 #!/usr/bin/env python3
-import os
-import zipfile
+"""
+zip_for_llms.py
+Package a repo for LLM consumption (zip and/or consolidated text), with optional
+Gemini CLI analysis over a filtered workspace and (optional) recent commit history.
+
+Key improvements in this release:
+- Verbose skip output is now compact: if a directory is excluded, we DO NOT print
+  any of its children. One line per excluded dir, period.
+- Restored readable colored output for humans (auto-disables when not a TTY or NO_COLOR set).
+- Gemini (-G) now shows a spinner/progress line while the CLI runs so you can
+  tell it's working instead of "hanging".
+- Presets expanded to mirror GitHub .gitignore templates for popular stacks
+  (python, cpp, node, rust, go, java, dotnet, swift). Using a preset adjusts
+  dirs/exts/files/remove-patterns by default.
+- NEW: Size tree reporting at the end of the run. Use:
+    -T / --size-tree [DEPTH]           (default 1 if value omitted)
+    -A / --auto-size-tree [TOPN]       (default 25 if value omitted; wins if both given)
+
+CLI style (single-letter + long-form, all programs must honor both):
+  -f / --file-mode           -> emit consolidated .txt
+  -z / --zip-mode            -> emit .zip
+  -o / --output              -> output base name/path (extension inferred if needed)
+  -x / --exclude-dir         -> exclude directory (repeatable)
+  -e / --exclude-ext         -> exclude extension (repeatable, with dot)
+  -X / --exclude-file        -> exclude filename (repeatable)
+  -I / --include-file        -> force-include filename (removes from exclude set)
+  -R / --remove-pattern      -> fnmatch to drop paths (repeatable)
+  -K / --keep-pattern        -> fnmatch to rescue paths (repeatable)
+  -m / --max-size            -> max zip size in MiB (best-effort culling)
+  -p / --preferences         -> comma list of preferred extensions to drop when trimming
+  -F / --flatten             -> copy files to a temp "_flattened" staging folder
+  -N / --name-by-path        -> if flattening, rename files by path (a_b_c.ext)
+  -v / --verbose             -> show progress & skip details
+  -P / --preset              -> use language preset (python, cpp, node, rust, go, java, dotnet, swift)
+  -G / --gemini              -> run Gemini CLI analysis on filtered workspace
+  -M / --model               -> Gemini model (default: gemini-2.5-flash)
+  -C / --include-commits     -> include a compact git commit snapshot in analysis workspace
+  -L / --commit-limit        -> limit the number of commits if included (default 20)
+  -S / --show-memory         -> add '--show-memory-usage' to Gemini CLI
+  -T / --size-tree [DEPTH]   -> print size tree to given depth (default 1 if no value)
+  -A / --auto-size-tree [N]  -> print "top N" largest folders (default 25). If deeper, parents included.
+
+The module also exposes:
+- zip_folder, text_file_mode, flatten_directory, delete_files_to_fit_size
+- prepare_analysis_workspace, run_gemini_analysis
+- compute_filtered_dir_sizes, render_size_tree
+- DEFAULT_EXCLUDE_DIRS/EXTS/FILES, PRESETS
+"""
+
+from __future__ import annotations
+
 import argparse
 import fnmatch
+import io
+import os
 import shutil
-from pathlib import Path
+import subprocess
+import sys
+import tempfile
+import threading
 import time
-
-try:
-    from rich.console import Console
-    from rich.text import Text
-    RICH_AVAILABLE = True
-except ImportError:
-    RICH_AVAILABLE = False
-
-# Fallback basic print if rich is not available
-if RICH_AVAILABLE:
-    console = Console()
-    # Define a generic print function that uses rich if available
-    def rprint(*args, **kwargs):
-        return console.print(*args, **kwargs)
-else:
-    # Basic print if rich is not available
-    def rprint(*args, **kwargs):
-        # Emulate rich's style argument roughly for simple cases if needed, or just print
-        if 'style' in kwargs:
-            del kwargs['style'] # Basic print doesn't understand 'style'
-        print(*args, **kwargs)
-    # Define Text as a no-op or simple string concat if rich is not available
-    class Text:
-        def __init__(self, text=""):
-            self._text = str(text)
-        def append(self, text, style=None): # style is ignored
-            self._text += str(text)
-        def __str__(self):
-            return self._text
+import zipfile
+from dataclasses import dataclass
+from pathlib import Path, PurePosixPath
+from typing import Dict, Iterable, List, Optional, Set, Tuple
 
 
-# Default patterns
-DEFAULT_EXCLUDE_DIRS = {
-    ".git", ".hg", ".svn", "__pycache__",
-    "node_modules", ".idea", ".vscode",
-    ".mypy_cache", ".pytest_cache"
-}
-DEFAULT_EXCLUDE_EXTS = {
-    ".pyc", ".pyo", ".exe", ".dll", ".bin",
-    ".o", ".so", ".zip", ".tar", ".gz", ".7z"
-}
-DEFAULT_EXCLUDE_FILES = {
-    "package-lock.json", "yarn.lock", "Pipfile.lock"
+# =====================================================================================
+# ANSI color helpers (auto-disable for non-TTY or NO_COLOR)
+# =====================================================================================
+
+def _supports_color() -> bool:
+    if os.environ.get("NO_COLOR"):
+        return False
+    try:
+        return sys.stdout.isatty()
+    except Exception:
+        return False
+
+_COLOR = _supports_color()
+
+def C(code: str) -> str:
+    return f"\033[{code}m" if _COLOR else ""
+
+COL = {
+    "reset": C("0"),
+    "bold": C("1"),
+    "dim": C("2"),
+    "green": C("32"),
+    "yellow": C("33"),
+    "red": C("31"),
+    "cyan": C("36"),
+    "magenta": C("35"),
+    "blue": C("34"),
+    "gray": C("90"),
 }
 
-def get_directory_size(directory: Path) -> int:
-    """Returns total size (bytes) of files under `directory`."""
-    return sum(
-        f.stat().st_size
-        for f in directory.rglob('*')
-        if f.is_file()
-    )
 
-def should_exclude(path: Path,
-                   exclude_dirs: set[str],
-                   exclude_exts: set[str],
-                   exclude_files: set[str]) -> bool:
-    if path.is_dir():
-        return path.name in exclude_dirs
-    return (
-        (path.suffix in exclude_exts) or
-        (path.name in exclude_files)
-    )
+def cfmt(s: str, color_key: str) -> str:
+    return f"{COL.get(color_key,'')}{s}{COL['reset'] if _COLOR else ''}"
 
-def matches_patterns(path_name: str, patterns: list[str]) -> bool:
-    return any(fnmatch.fnmatch(path_name, pat) for pat in patterns)
+
+# =====================================================================================
+# Defaults & Extended Presets
+# =====================================================================================
+
+DEFAULT_EXCLUDE_DIRS: Set[str] = {
+    # VCS & IDE
+    ".git", ".hg", ".svn", ".idea", ".vscode", ".DS_Store",
+    # Python caches / envs
+    "__pycache__", ".mypy_cache", ".pytest_cache", ".venv", "env", "venv", ".tox",
+    # Node
+    "node_modules",
+    # Build / dist
+    "dist", "build", "target", "out",
+}
+
+DEFAULT_EXCLUDE_EXTS: Set[str] = {
+    # compiled
+    ".pyc", ".pyo", ".pyd", ".o", ".obj", ".class",
+    # archives & binaries
+    ".zip", ".tar", ".gz", ".7z", ".rar", ".iso",
+    ".dll", ".so", ".dylib", ".exe",
+    # media-ish large odds
+    ".jpg", ".jpeg", ".png", ".gif", ".bmp", ".tiff", ".mp4", ".mov", ".mp3", ".wav",
+    # package locks often noisy
+    ".lock",
+}
+
+DEFAULT_EXCLUDE_FILES: Set[str] = {
+    "package-lock.json", "yarn.lock", "pnpm-lock.yaml",
+    ".coverage", "coverage.xml",
+}
+
+# Hefty, language-leaning presets inspired by github/gitignore contents.
+# Note: these are static patterns; they don’t fetch from the web.
+PRESETS = {
+    "python": {
+        "dirs": {
+            ".venv", "venv", "env", "__pycache__", ".pytest_cache",
+            ".mypy_cache", ".tox", "build", "dist", "*.egg-info",
+        },
+        "exts": {".pyc", ".pyo", ".pyd", ".so"},
+        "files": {".python-version", ".coverage", "coverage.xml"},
+        "patterns": ["*.egg-info", ".ruff_cache", ".hypothesis", ".nox"],
+    },
+    "cpp": {
+        "dirs": {"build", "cmake-build-*", "out", "bin", "obj", ".cache"},
+        "exts": {".o", ".obj", ".a", ".lib", ".dll", ".so", ".dylib", ".exe", ".pdb"},
+        "files": set(),
+        "patterns": ["CMakeFiles", "CMakeCache.txt", "compile_commands.json"],
+    },
+    "node": {
+        "dirs": {"node_modules", "dist", "coverage", ".next", ".nuxt", ".turbo"},
+        "exts": {".log", ".map"},
+        "files": {"package-lock.json", "yarn.lock", "pnpm-lock.yaml"},
+        "patterns": [".vercel", ".cache", ".parcel-cache"],
+    },
+    "rust": {
+        "dirs": {"target"},
+        "exts": {".rlib", ".d", ".o"},
+        "files": set(),
+        "patterns": ["Cargo.lock"],
+    },
+    "go": {
+        "dirs": {"bin"},
+        "exts": {".test"},
+        "files": {"go.sum"},
+        "patterns": ["vendor", ".cache"],
+    },
+    "java": {
+        "dirs": {"target", "out", ".gradle", ".idea"},
+        "exts": {".class", ".jar", ".war", ".ear"},
+        "files": {"pom.xml.tag"},
+        "patterns": ["*.iml", "build"],
+    },
+    "dotnet": {
+        "dirs": {"bin", "obj", ".vs"},
+        "exts": {".dll", ".pdb", ".cache", ".mdb"},
+        "files": {"project.lock.json"},
+        "patterns": ["*.user", "*.suo"],
+    },
+    "swift": {
+        "dirs": {".build", "DerivedData"},
+        "exts": {".xcodeproj", ".xcworkspace"},
+        "files": set(),
+        "patterns": ["*.swiftpm"],
+    },
+}
+
+# =====================================================================================
+# Tiny structured printing utilities
+# =====================================================================================
+
+def _print_header(title: str) -> None:
+    print(f"{cfmt('==', 'magenta')} {cfmt(title, 'bold')}")
+
+def _print_stat(label: str, value: str) -> None:
+    print(f"  {cfmt(label+':', 'cyan')} {value}")
+
+def _print_skip(path: str, why: str) -> None:
+    print(f"  - {cfmt(path, 'gray')} {cfmt(f'({why})', 'yellow')}")
+
+def _print_ok(path: str) -> None:
+    print(f"  + {cfmt(path, 'green')}")
+
+def _print_warn(msg: str) -> None:
+    print(f"{cfmt('warn:', 'yellow')} {msg}")
+
+def _print_err(msg: str) -> None:
+    print(f"{cfmt('error:', 'red')} {msg}", file=sys.stderr)
+
+
+# =====================================================================================
+# Helpers
+# =====================================================================================
+
+def _relpath(p: Path, root: Path) -> str:
+    try:
+        return str(PurePosixPath(p.relative_to(root)))
+    except Exception:
+        return str(PurePosixPath(p))
+
+
+def _match_any_fragment(path_rel: str, patterns: Iterable[str]) -> bool:
+    # Consider any fragment match: exact dir name, partial path fragment, or glob
+    for pat in patterns:
+        if pat in path_rel:
+            return True
+        if fnmatch.fnmatch(path_rel, pat):
+            return True
+    return False
+
 
 def flatten_directory(source_dir: Path, name_by_path: bool, verbose: bool) -> Path:
-    source_dir = source_dir.resolve()
-    flat_dir_name = "_flattened"
-    flat_dest_path = source_dir / flat_dir_name
-    flat_dest_path.mkdir(exist_ok=True)
-    moved_count = 0
-
-    if verbose:
-        rprint(f"Flattening directory [cyan]{source_dir.name}[/cyan] into [cyan]{flat_dest_path.name}[/cyan]...")
-
-    for root, dirs, files in os.walk(source_dir, topdown=True):
+    """
+    Copy all files into a temp "_flattened" directory for packaging.
+    """
+    flat = source_dir / "_flattened"
+    if flat.exists():
+        shutil.rmtree(flat)
+    flat.mkdir()
+    for root, _, files in os.walk(source_dir):
         root_path = Path(root)
-
-        if root_path == flat_dest_path:
-            dirs[:] = [] 
+        if flat in root_path.parents or root_path == flat:
             continue
+        for f in files:
+            src = root_path / f
+            rel = _relpath(src, source_dir)
+            target_name = rel.replace("/", "_") if name_by_path else f
+            dst = flat / target_name
+            if verbose:
+                _print_ok(f"flatten: {rel} -> {_relpath(dst, source_dir)}")
+            try:
+                shutil.copy2(src, dst)
+            except Exception as e:
+                if verbose:
+                    _print_warn(f"failed to copy {rel}: {e}")
+    # Add a plain structure marker file for zip output parity
+    (flat / "folder_structure.txt").write_text("Flattened view artifacts (see text mode for full structure).")
+    return flat
 
-        dirs[:] = [d for d in dirs if d not in DEFAULT_EXCLUDE_DIRS and d != flat_dir_name]
 
-        if root_path.name in DEFAULT_EXCLUDE_DIRS and root_path != source_dir:
-            continue
+def delete_files_to_fit_size(folder: Path, max_mib: int, preference_exts: Iterable[str], verbose: bool) -> List[str]:
+    """
+    Best-effort removal: delete biggest preferred-extension files first until
+    the folder fits under max_mib.
+    """
+    max_bytes = max_mib * 1024 * 1024
+    removed: List[str] = []
 
-        for fn in files:
-            original_file_path = root_path / fn
-            
-            if should_exclude(original_file_path, set(), DEFAULT_EXCLUDE_EXTS, DEFAULT_EXCLUDE_FILES):
-                if verbose: rprint(f"  [dim yellow]Skipping (default rule):[/dim yellow] {original_file_path.relative_to(source_dir)}")
-                continue
-            
-            if original_file_path.parent == flat_dest_path:
-                continue
-
-            if name_by_path:
+    def size_of_tree(p: Path) -> int:
+        total = 0
+        for r, _, files in os.walk(p):
+            for f in files:
                 try:
-                    relative_path_from_source = original_file_path.relative_to(source_dir)
-                    new_name = "_".join(relative_path_from_source.parts)
-                except ValueError: 
-                    new_name = fn 
-            else:
-                new_name = fn
-            
-            destination_file_path_in_flat = flat_dest_path / new_name
-            
-            if destination_file_path_in_flat != original_file_path:
-                shutil.move(str(original_file_path), str(destination_file_path_in_flat))
-                moved_count +=1
-                # if verbose: rprint(f"  [dim]Moved:[/dim] {original_file_path.name} -> {new_name}")
+                    total += (Path(r) / f).stat().st_size
+                except OSError:
+                    pass
+        return total
 
-
-    rprint(f"[green]Flattened {moved_count} files into {flat_dest_path}.[/green]")
-    return flat_dest_path
-
-def delete_files_to_fit_size(directory_to_prune: Path,
-                             target_zip_size_mb: int, 
-                             preferences: list[str],
-                             verbose: bool) -> list[str]:
-    target_dir_bytes = target_zip_size_mb * 1024 * 1024 
-    current_dir_size = get_directory_size(directory_to_prune)
-
-    if current_dir_size <= target_dir_bytes : 
-        if verbose: rprint(f"Directory '{directory_to_prune}' size ({current_dir_size/(1024*1024):.2f}MB) is already below/near heuristic target ({target_zip_size_mb}MB). No files deleted by pruning logic itself.", style="dim")
-        return []
-
-    if verbose: rprint(f"Attempting to reduce source directory [cyan]'{directory_to_prune.name}'[/cyan] size... Current: {current_dir_size/(1024*1024):.2f} MB. Heuristic target: {target_zip_size_mb} MB")
-    
-    all_files_in_dir = sorted(
-        (f for f in directory_to_prune.rglob('*') if f.is_file()),
-        key=lambda f: f.stat().st_size,
-        reverse=True
-    )
-
-    ordered_files_for_deletion: list[Path] = []
-    for pref_ext in preferences:
-        ordered_files_for_deletion.extend([f for f in all_files_in_dir if f.suffix == pref_ext and f not in ordered_files_for_deletion])
-    ordered_files_for_deletion.extend([f for f in all_files_in_dir if f not in ordered_files_for_deletion])
-
-    removed_paths = []
-    if verbose and ordered_files_for_deletion:
-         rprint("  [bold yellow]Pruning files:[/bold yellow]")
-
-    for file_to_delete in ordered_files_for_deletion:
-        if current_dir_size <= target_dir_bytes:
-            break
-        try:
-            file_size = file_to_delete.stat().st_size
-            file_to_delete.unlink() 
-            removed_paths.append(str(file_to_delete))
-            current_dir_size -= file_size
-            if verbose: rprint(f"    - [red]Pruned:[/red] {file_to_delete.name} ({file_size/(1024*1024):.2f}MB). Source dir size now: {current_dir_size/(1024*1024):.2f}MB")
-        except OSError as e:
-            rprint(f"  [yellow]Warning:[/yellow] Could not delete file {file_to_delete} during pruning: {e}")
-
-    if current_dir_size <= target_dir_bytes:
-        if verbose: rprint(f"  [green]Pruned {len(removed_paths)} files.[/green] Final source dir size for '{directory_to_prune.name}': {current_dir_size/(1024*1024):.2f} MB.")
-    else:
-        if verbose: rprint(f"  [yellow]Warning:[/yellow] After pruning {len(removed_paths)} files, source dir size for '{directory_to_prune.name}' ({current_dir_size/(1024*1024):.2f}MB) may still be too large for heuristic target ({target_zip_size_mb}MB).")
-    return removed_paths
-
-
-def zip_folder(source_dir_str: str,
-               output_zip_str: str,
-               exclude_dirs: set[str],
-               exclude_exts: set[str],
-               exclude_files: set[str],
-               remove_patterns: list[str],
-               keep_patterns: list[str],
-               max_size: int | None,
-               preferences: list[str],
-               flatten: bool,
-               name_by_path: bool,
-               verbose: bool) -> None:
-    src_path = Path(source_dir_str).resolve()
-    output_zip_path = Path(output_zip_str).resolve()
-    
-    temp_source_copy_for_pruning: Path | None = None
-    temp_dir_to_flatten_base: Path | None = None 
-
-    if flatten:
-        temp_dir_to_flatten_base = src_path.parent / f"_temp_to_flatten_{src_path.name}_{int(time.time())}"
-        shutil.copytree(src_path, temp_dir_to_flatten_base, dirs_exist_ok=True)
-        working_dir_for_zip = flatten_directory(temp_dir_to_flatten_base, name_by_path, verbose) 
-        if verbose: rprint(f"Flatten mode: Zipping from temporary flattened directory: [cyan]{working_dir_for_zip}[/cyan]")
-    elif not flatten and max_size is not None: 
-        temp_source_copy_for_pruning = src_path.parent / f"_temp_zip_src_{src_path.name}_{int(time.time())}"
-        if verbose: rprint(f"Max size set and not flattening. Using temp copy for zipping: [cyan]{temp_source_copy_for_pruning}[/cyan]")
-        shutil.copytree(src_path, temp_source_copy_for_pruning, dirs_exist_ok=True)
-        working_dir_for_zip = temp_source_copy_for_pruning
-    else: 
-        working_dir_for_zip = src_path
-    
-    if not working_dir_for_zip.exists():
-        rprint(f"[red]Error:[/red] Working directory '{working_dir_for_zip}' does not exist. Aborting zip.")
-        if temp_dir_to_flatten_base and temp_dir_to_flatten_base.exists(): shutil.rmtree(temp_dir_to_flatten_base)
-        if temp_source_copy_for_pruning and temp_source_copy_for_pruning.exists(): shutil.rmtree(temp_source_copy_for_pruning)
-        return
-
-    hierarchy_file_path = working_dir_for_zip / "folder_structure.txt"
-    with open(hierarchy_file_path, 'w', encoding='utf-8') as f:
-        f.write(f"Folder Structure for: {src_path.name}\n")
-        paths_for_structure = []
-        for root, dirs, _ in os.walk(src_path, topdown=True):
-            current_root_path = Path(root)
-            dirs[:] = [
-                d_name for d_name in dirs
-                if not (
-                    should_exclude(current_root_path / d_name, exclude_dirs, set(), set()) or
-                    (matches_patterns(d_name, remove_patterns) and not matches_patterns(d_name, keep_patterns))
-                )
-            ]
-            relative_path = current_root_path.relative_to(src_path)
-            indent_level = len(relative_path.parts)
-            indent = '    ' * indent_level
-            dir_name_display = current_root_path.name if current_root_path != src_path else src_path.name
-            paths_for_structure.append(f"{indent}{dir_name_display}/\n")
-        for line in paths_for_structure:
-            f.write(line)
-    if verbose:
-        rprint(f"Generated [magenta]folder_structure.txt[/magenta] at [cyan]{hierarchy_file_path}[/cyan]")
-
+    preference_exts = set(preference_exts or [])
     while True:
-        zip_file_count = 0
-        output_zip_path.parent.mkdir(parents=True, exist_ok=True)
-        
-        if verbose:
-            rprint(f"\n[bold blue]Processing files for zip '{output_zip_path.name}':[/bold blue]")
-
-        with zipfile.ZipFile(output_zip_path, "w", zipfile.ZIP_DEFLATED) as zf:
-            for root, dirs, files_in_dir in os.walk(working_dir_for_zip, topdown=True):
-                current_root_path_in_working = Path(root)
-                dirs[:] = [
-                    d_name for d_name in dirs
-                    if not (
-                        should_exclude(current_root_path_in_working / d_name, exclude_dirs, set(), set()) or
-                        (matches_patterns(d_name, remove_patterns) and not matches_patterns(d_name, keep_patterns))
-                    )
-                ]
-                for filename in files_in_dir:
-                    file_path_in_working = current_root_path_in_working / filename
-                    relative_display_path = file_path_in_working.relative_to(working_dir_for_zip)
-                    
-                    if should_exclude(file_path_in_working, exclude_dirs, exclude_exts, exclude_files):
-                        if verbose: rprint(f"  - [yellow]Skipping (default rule):[/yellow] {relative_display_path}")
-                        continue
-                    
-                    if matches_patterns(filename, remove_patterns) and not matches_patterns(filename, keep_patterns):
-                        if verbose: rprint(f"  - [yellow]Skipping (pattern rule on '{filename}'):[/yellow] {relative_display_path}")
-                        continue
-                    
-                    archive_name = relative_display_path
-                    zf.write(file_path_in_working, archive_name)
-                    zip_file_count += 1
-                    if verbose: 
-                        rprint(f"  + [green]Added to zip:[/green] {archive_name}")
-        
-        if verbose:
-            rprint(f"Total files added to current zip iteration: {zip_file_count}")
-
-        current_zip_size_bytes = output_zip_path.stat().st_size
-        current_zip_size_mb = current_zip_size_bytes / (1024 * 1024)
-
-        if max_size is None or current_zip_size_mb <= max_size:
-            rprint(f"\n[bold green]Final zip file '{output_zip_path.name}' size: {current_zip_size_mb:.2f} MB[/bold green]")
+        total = size_of_tree(folder)
+        if total <= max_bytes:
             break
-        
-        if verbose: rprint(f"\n[bold yellow]Zip file is too large ({current_zip_size_mb:.2f} MB > {max_size} MB).[/bold yellow] Pruning files from [cyan]'{working_dir_for_zip.name}'[/cyan]...")
-        
-        deleted_files = delete_files_to_fit_size(working_dir_for_zip, max_size, preferences, verbose)
-        
-        if not deleted_files:
-            rprint(f"[yellow]Warning:[/yellow] No files were deleted from '{working_dir_for_zip.name}' during pruning, but zip is still too large ({current_zip_size_mb:.2f} MB > {max_size} MB). "
-                  "The zip might exceed max_size if no more files can be pruned or preferences don't match. Check file sizes and compressibility.")
-            break 
-
-    if temp_source_copy_for_pruning and temp_source_copy_for_pruning.exists():
-        if verbose: rprint(f"Cleaning up temporary source copy for zipping: [cyan]{temp_source_copy_for_pruning}[/cyan]", style="dim")
-        shutil.rmtree(temp_source_copy_for_pruning)
-    if temp_dir_to_flatten_base and temp_dir_to_flatten_base.exists():
-        if verbose: rprint(f"Cleaning up temporary flattened structure for zipping: [cyan]{temp_dir_to_flatten_base}[/cyan]", style="dim")
-        shutil.rmtree(temp_dir_to_flatten_base)
-
-
-def text_file_mode(source_dir_str: str,
-                   output_file_str: str,
-                   exclude_dirs: set[str],
-                   exclude_exts: set[str],
-                   exclude_files: set[str],
-                   remove_patterns: list[str],
-                   keep_patterns: list[str],
-                   flatten_mode: bool, 
-                   name_by_path: bool,
-                   verbose: bool) -> None:
-    src_path = Path(source_dir_str).resolve()
-    output_file_path = Path(output_file_str).resolve()
-
-    temp_flatten_root_for_text: Path | None = None
-    working_dir_for_text = src_path
-
-    if flatten_mode:
-        temp_flatten_root_for_text = src_path.parent / f"_temp_text_flatten_base_{src_path.name}_{int(time.time())}"
-        shutil.copytree(src_path, temp_flatten_root_for_text, dirs_exist_ok=True)
-        working_dir_for_text = flatten_directory(temp_flatten_root_for_text, name_by_path, verbose)
-        if verbose: rprint(f"Text mode: Using temporary flattened content from: [cyan]{working_dir_for_text}[/cyan]")
-
-    files_processed_count = 0
-    files_added_to_text_count = 0
-    files_skipped_list: list[str] = []
-    paths_for_hierarchy_display: list[str] = []
-    
-    output_file_path.parent.mkdir(parents=True, exist_ok=True)
-    with open(output_file_path, 'w', encoding='utf-8') as out_f:
-        rprint(f"[bold blue]Generating folder structure for text output...[/bold blue]")
-        out_f.write(f"Folder Structure for: {src_path.name}\n") 
-        for root, dirs, _ in os.walk(src_path, topdown=True): 
-            current_root_path = Path(root)
-            dirs[:] = [
-                d_name for d_name in dirs
-                if not (
-                    should_exclude(current_root_path / d_name, exclude_dirs, set(), set()) or
-                    (matches_patterns(d_name, remove_patterns) and not matches_patterns(d_name, keep_patterns))
-                )
-            ]
-            relative_path = current_root_path.relative_to(src_path)
-            indent_level = len(relative_path.parts)
-            indent = '    ' * indent_level
-            dir_name_display = current_root_path.name if current_root_path != src_path else src_path.name
-            paths_for_hierarchy_display.append(f"{indent}{dir_name_display}/\n")
-        for line in paths_for_hierarchy_display:
-            out_f.write(line)
-        out_f.write("\n--- End of Folder Structure ---\n")
-        if verbose: rprint("Folder structure written to text file.")
-
-
-        if verbose:
-            rprint(f"\n[bold blue]Processing files for text output '{output_file_path.name}':[/bold blue]")
-
-        for root, dirs, files_in_dir in os.walk(working_dir_for_text, topdown=True):
-            current_root_path_in_working = Path(root)
-            dirs[:] = [
-                d_name for d_name in dirs
-                if not (
-                    should_exclude(current_root_path_in_working / d_name, exclude_dirs, set(), set()) or
-                    (matches_patterns(d_name, remove_patterns) and not matches_patterns(d_name, keep_patterns))
-                )
-            ]
-
-            for filename in files_in_dir:
-                files_processed_count += 1
-                file_path_in_working = current_root_path_in_working / filename
-                
-                if flatten_mode:
-                    rel_path_for_display_str = filename
-                else: 
-                    try:
-                        rel_path_for_display_str = str(file_path_in_working.relative_to(src_path))
-                    except ValueError: 
-                        rel_path_for_display_str = filename
-
-                if should_exclude(file_path_in_working, exclude_dirs, exclude_exts, exclude_files):
-                    skip_msg = f"{rel_path_for_display_str} (excluded by dir/ext/file rule on '{file_path_in_working.name}')"
-                    files_skipped_list.append(skip_msg)
-                    if verbose: rprint(f"  - [yellow]Skipping (default rule):[/yellow] {skip_msg}")
-                    continue
-                if matches_patterns(filename, remove_patterns) and not matches_patterns(filename, keep_patterns):
-                    skip_msg = f"{rel_path_for_display_str} (pattern on name '{filename}')"
-                    files_skipped_list.append(skip_msg)
-                    if verbose: rprint(f"  - [yellow]Skipping (pattern rule):[/yellow] {skip_msg}")
-                    continue
-
-                out_f.write(f"\n-- File: {rel_path_for_display_str} --\n")
+        # collect candidate files
+        candidates: List[Path] = []
+        for r, _, files in os.walk(folder):
+            for f in files:
+                p = Path(r) / f
                 try:
-                    content = file_path_in_working.read_text(encoding='utf-8')
-                    out_f.write(content)
-                    if not content.endswith('\n'):
-                        out_f.write('\n')
-                    files_added_to_text_count += 1
-                    if verbose: 
-                        rprint(f"  + [green]Added to text:[/green] {rel_path_for_display_str}")
-                except UnicodeDecodeError:
-                    skip_msg = f"{rel_path_for_display_str} (binary or non-UTF-8 content)"
-                    files_skipped_list.append(skip_msg)
-                    if verbose: rprint(f"  - [yellow]Skipping (read error):[/yellow] {skip_msg}")
-                except Exception as e:
-                    skip_msg = f"{rel_path_for_display_str} (read error: {e})"
-                    files_skipped_list.append(skip_msg)
-                    if verbose: rprint(f"  - [yellow]Skipping (read error):[/yellow] {skip_msg}")
-        
-    if temp_flatten_root_for_text and temp_flatten_root_for_text.exists():
+                    if p.suffix in preference_exts:
+                        candidates.append(p)
+                except Exception:
+                    continue
+        if not candidates:
+            # fallback: delete any biggest file
+            for r, _, files in os.walk(folder):
+                for f in files:
+                    candidates.append(Path(r) / f)
+        if not candidates:
+            break
+        # pick largest
+        candidates.sort(key=lambda p: p.stat().st_size if p.exists() else 0, reverse=True)
+        victim = candidates[0]
         try:
-            shutil.rmtree(temp_flatten_root_for_text)
-            if verbose: rprint(f"Cleaned up temporary flatten base for text mode: [cyan]{temp_flatten_root_for_text}[/cyan]", style="dim")
-        except OSError as e:
-            rprint(f"[yellow]Warning:[/yellow] Could not clean up temporary directory {temp_flatten_root_for_text}: {e}")
+            if verbose:
+                _print_warn(f"trimming to fit: deleting {_relpath(victim, folder)}")
+            removed.append(str(victim))
+            victim.unlink(missing_ok=True)
+        except Exception as e:
+            if verbose:
+                _print_warn(f"failed to delete {victim}: {e}")
+            break
+    return removed
 
-    rprint(f"\n[bold green]Text file generation complete: {output_file_path.name}[/bold green]")
-    rprint(f"Total files encountered in content source: {files_processed_count}")
-    rprint(f"Files added to text output: {files_added_to_text_count}")
-    if files_skipped_list:
-        rprint(f"Files skipped from content: {len(files_skipped_list)}") 
-        if verbose: # List skipped files only in verbose mode for cleaner default output
-            rprint("  [bold yellow]Skipped file details:[/bold yellow]")
-            for f_skipped in files_skipped_list:
-                rprint(f"    - {f_skipped}")
 
-if __name__ == "__main__":
-    if not RICH_AVAILABLE:
-        print("Warning: 'rich' library not found. Output will be basic.")
-        print("Consider installing it with: pip install rich\n")
+# =====================================================================================
+# Exclusion logic
+# =====================================================================================
 
-    parser = argparse.ArgumentParser(
-        description="Package a folder for LLMs (zip and/or text file mode)."
+@dataclass(frozen=True)
+class Exclusion:
+    exclude_dirs: Set[str]
+    exclude_exts: Set[str]
+    exclude_files: Set[str]
+    remove_patterns: List[str]
+    keep_patterns: List[str]
+
+    def dir_is_excluded(self, rel_dir: str) -> bool:
+        # If directory name itself or any fragment/glob matches, it's excluded
+        name = Path(rel_dir).name
+        if name in self.exclude_dirs:
+            return True
+        if rel_dir in self.exclude_dirs:
+            return True
+        return _match_any_fragment(rel_dir, self.exclude_dirs)
+
+    def file_is_excluded(self, rel_path: str) -> bool:
+        # Keep overrides remove
+        base = Path(rel_path).name
+        if base in self.keep_patterns or any(fnmatch.fnmatch(base, pat) for pat in self.keep_patterns):
+            return False
+        if base in self.exclude_files:
+            return True
+        if Path(rel_path).suffix in self.exclude_exts:
+            return True
+        if any(fnmatch.fnmatch(rel_path, pat) for pat in self.remove_patterns):
+            return True
+        # If any parent dir is excluded by pattern, treat as excluded (pruning handles this too)
+        parts = Path(rel_path).parts
+        agg = ""
+        for part in parts[:-1]:
+            agg = f"{agg}/{part}" if agg else part
+            if self.dir_is_excluded(agg):
+                return True
+        return False
+
+
+# =====================================================================================
+# Text mode
+# =====================================================================================
+
+def _emit_folder_structure(root: Path, exclusion: Exclusion, out: io.TextIOBase, verbose: bool) -> None:
+    out.write(f"\nFolder Structure for: {root.name}\n")
+    def walk(d: Path, indent: int = 0) -> None:
+        rel = _relpath(d, root)
+        if rel and exclusion.dir_is_excluded(rel):
+            out.write("    " * indent + f"{Path(rel).name}/ (excluded)\n")
+            if verbose:
+                _print_skip(rel, "excluded dir")
+            return  # CRITICAL: do not descend into excluded dirs
+        if rel:
+            out.write("    " * indent + f"{Path(rel).name}/\n")
+        for child in sorted(d.iterdir(), key=lambda p: (not p.is_dir(), p.name.lower())):
+            if child.is_dir():
+                walk(child, indent + (1 if rel else 1))
+    walk(root)
+
+
+def text_file_mode(
+    source_dir_str: str,
+    output_txt_str: str,
+    exclude_dirs: Set[str],
+    exclude_exts: Set[str],
+    exclude_files: Set[str],
+    remove_patterns: List[str],
+    keep_patterns: List[str],
+    flatten: bool,
+    name_by_path: bool,
+    verbose: bool,
+) -> None:
+    source_dir = Path(source_dir_str).resolve()
+    output_txt = Path(output_txt_str)
+
+    exclusion = Exclusion(set(exclude_dirs), set(exclude_exts), set(exclude_files), list(remove_patterns), list(keep_patterns))
+
+    if verbose:
+        _print_header("Generating folder structure...")
+    with io.open(output_txt, "w", encoding="utf-8") as out:
+        out.write("This document packages a repository for a Large Language Model (LLM).\n")
+        _emit_folder_structure(source_dir, exclusion, out, verbose)
+
+        out.write("\nProcessing files for content...\n")
+
+        skipped_once_dirs: Set[str] = set()  # remember which excluded dirs we already logged
+
+        # Prepare flattening if requested (still prints structure from original tree)
+        staging_root = source_dir
+        if flatten:
+            if verbose:
+                _print_header("Flattening files...")
+            staging_root = flatten_directory(source_dir, name_by_path=name_by_path, verbose=verbose)
+
+        # Walk staging root for file contents (prune excluded dirs entirely)
+        for r, dirs, files in os.walk(staging_root):
+            root_path = Path(r)
+            # Prune excluded dirs BEFORE descent; print just once per dir.
+            pruned: List[str] = []
+            for d in list(dirs):
+                rel_d = _relpath(root_path / d, staging_root)
+                if exclusion.dir_is_excluded(rel_d):
+                    pruned.append(d)
+                    if verbose and rel_d not in skipped_once_dirs:
+                        _print_skip(rel_d, "excluded dir")
+                        skipped_once_dirs.add(rel_d)
+            for d in pruned:
+                dirs.remove(d)
+
+            # Process files
+            for f in files:
+                p = root_path / f
+                rel = _relpath(p, staging_root)
+                if exclusion.file_is_excluded(rel):
+                    # Only print skip if not under an already-excluded dir (we pruned, but be safe)
+                    parent_rel = _relpath(p.parent, staging_root)
+                    if verbose and not exclusion.dir_is_excluded(parent_rel):
+                        reason = "excluded" if Path(rel).suffix in exclusion.exclude_exts or Path(rel).name in exclusion.exclude_files else "removed by pattern"
+                        _print_skip(rel, reason)
+                    continue
+
+                # Emit header then try to read as UTF-8 text
+                display_name = rel.replace("/", "_") if (flatten and name_by_path) else rel
+                out.write(f"\n-- File: {display_name} --\n")
+                try:
+                    with io.open(p, "r", encoding="utf-8") as fin:
+                        out.write(fin.read())
+                except UnicodeDecodeError:
+                    if verbose:
+                        _print_skip(rel, "binary or non-UTF-8 content")
+                except Exception as e:
+                    if verbose:
+                        _print_warn(f"read error {rel}: {e}")
+
+        if flatten and staging_root != source_dir:
+            shutil.rmtree(staging_root, ignore_errors=True)
+
+    if verbose:
+        _print_stat("Text file complete", str(output_txt))
+        _print_stat("Source", str(source_dir))
+
+
+# =====================================================================================
+# Zip mode
+# =====================================================================================
+
+def zip_folder(
+    source_dir_str: str,
+    output_zip_str: str,
+    exclude_dirs: Set[str],
+    exclude_exts: Set[str],
+    exclude_files: Set[str],
+    remove_patterns: List[str],
+    keep_patterns: List[str],
+    max_size: Optional[int],
+    preferences: List[str],
+    flatten: bool,
+    name_by_path: bool,
+    verbose: bool,
+) -> None:
+    source_dir = Path(source_dir_str).resolve()
+    output_zip = Path(output_zip_str)
+    exclusion = Exclusion(set(exclude_dirs), set(exclude_exts), set(exclude_files), list(remove_patterns), list(keep_patterns))
+
+    # Stage files (optionally flatten)
+    staging_root = source_dir
+    if flatten:
+        if verbose:
+            _print_header("Flattening files...")
+        staging_root = flatten_directory(source_dir, name_by_path=name_by_path, verbose=verbose)
+
+    # Create zip
+    if verbose:
+        _print_header(f"Creating zip: {output_zip.name}")
+    with zipfile.ZipFile(output_zip, "w", compression=zipfile.ZIP_DEFLATED) as z:
+        # Add a structure marker (helps LLM context)
+        structure_txt = "folder_structure.txt"
+        z.writestr(structure_txt, "See text mode for detailed structure.")
+        if verbose:
+            _print_ok(f"added {structure_txt}")
+
+        for r, dirs, files in os.walk(staging_root):
+            root_path = Path(r)
+            # Prune excluded dirs before descent
+            for d in list(dirs):
+                rel_d = _relpath(root_path / d, staging_root)
+                if exclusion.dir_is_excluded(rel_d):
+                    if verbose:
+                        _print_skip(rel_d, "excluded dir")
+                    dirs.remove(d)
+            for f in files:
+                p = root_path / f
+                rel = _relpath(p, staging_root)
+                if exclusion.file_is_excluded(rel):
+                    if verbose:
+                        reason = "excluded"
+                        _print_skip(rel, reason)
+                    continue
+                arcname = rel.replace("/", "_") if (flatten and name_by_path) else rel
+                try:
+                    z.write(p, arcname)
+                    if verbose:
+                        _print_ok(f"+ {arcname}")
+                except Exception as e:
+                    if verbose:
+                        _print_warn(f"failed to add {rel}: {e}")
+
+    # Size trimming if requested
+    if max_size is not None:
+        # Extract, trim, re-zip (simpler logic)
+        with tempfile.TemporaryDirectory() as td:
+            td_path = Path(td)
+            with zipfile.ZipFile(output_zip, "r") as z:
+                z.extractall(td_path)
+            removed = delete_files_to_fit_size(td_path, max_size, preferences, verbose=verbose)
+            with zipfile.ZipFile(output_zip, "w", compression=zipfile.ZIP_DEFLATED) as z:
+                for r, _, files in os.walk(td_path):
+                    for f in files:
+                        p = Path(r) / f
+                        rel = _relpath(p, td_path)
+                        z.write(p, rel)
+            if verbose and removed:
+                _print_warn(f"trimmed {len(removed)} files to honor max size")
+
+    if flatten and staging_root != source_dir:
+        shutil.rmtree(staging_root, ignore_errors=True)
+
+    if verbose:
+        _print_stat("Zip complete", str(output_zip))
+
+
+# =====================================================================================
+# Gemini analysis
+# =====================================================================================
+
+def _spinner(stop_event: threading.Event, label: str = "Running Gemini") -> None:
+    frames = "|/-\\"
+    i = 0
+    start = time.time()
+    while not stop_event.is_set():
+        elapsed = int(time.time() - start)
+        msg = f"\r{cfmt(frames[i % len(frames)], 'cyan')} {label}… {elapsed}s"
+        sys.stdout.write(msg)
+        sys.stdout.flush()
+        time.sleep(0.15)
+        i += 1
+    # clear line
+    sys.stdout.write("\r" + " " * 60 + "\r")
+    sys.stdout.flush()
+
+
+def _git_commit_snapshot(repo_root: Path, limit: Optional[int]) -> str | None:
+    try:
+        cmd = ["git", "log", "--pretty=format:%h %ad %s", "--date=short"]
+        if limit:
+            cmd.extend(["-n", str(limit)])
+        cp = subprocess.run(cmd, cwd=str(repo_root), capture_output=True, text=True)
+        if cp.returncode == 0 and cp.stdout.strip():
+            return cp.stdout.strip()
+    except Exception:
+        pass
+    return None
+
+
+def prepare_analysis_workspace(
+    source_dir: Path,
+    exclude_dirs: Set[str],
+    exclude_exts: Set[str],
+    exclude_files: Set[str],
+    remove_patterns: List[str],
+    keep_patterns: List[str],
+    verbose: bool,
+) -> Path:
+    """
+    Copy filtered tree into a temporary workspace for Gemini CLI to analyze.
+    """
+    exclusion = Exclusion(set(exclude_dirs), set(exclude_exts), set(exclude_files), list(remove_patterns), list(keep_patterns))
+    ws = Path(tempfile.mkdtemp(prefix="llm_ws_"))
+
+    if verbose:
+        _print_header("Generating analysis workspace...")
+
+    for r, dirs, files in os.walk(source_dir):
+        root_path = Path(r)
+        rel_dir = _relpath(root_path, source_dir)
+        # Prune & mirror directories
+        pruned: List[str] = []
+        for d in list(dirs):
+            rel_d = _relpath(root_path / d, source_dir)
+            if exclusion.dir_is_excluded(rel_d):
+                pruned.append(d)
+                if verbose:
+                    _print_skip(rel_d, "excluded dir")
+        for d in pruned:
+            dirs.remove(d)
+        # Ensure mirrored directory exists in ws
+        target_dir = ws / (rel_dir if rel_dir != "." else "")
+        target_dir.mkdir(parents=True, exist_ok=True)
+        # Copy files
+        for f in files:
+            src = root_path / f
+            rel = _relpath(src, source_dir)
+            if exclusion.file_is_excluded(rel):
+                if verbose:
+                    _print_skip(rel, "excluded")
+                continue
+            dst = target_dir / f
+            try:
+                shutil.copy2(src, dst)
+            except Exception as e:
+                if verbose:
+                    _print_warn(f"copy failed {rel}: {e}")
+
+    return ws
+
+
+def run_gemini_analysis(
+    filtered_workspace: Path,
+    model_name: str,
+    analysis_outfile: Path,
+    include_commits: bool,
+    commit_limit: Optional[int],
+    show_memory: bool,
+    verbose: bool,
+) -> int:
+    """
+    Execute 'gemini' CLI with all-files flag and optional commit history.
+    Shows a spinner while the subprocess runs so the UI isn't "frozen".
+    """
+    if include_commits:
+        snap = _git_commit_snapshot(filtered_workspace, commit_limit)
+        if snap:
+            (filtered_workspace / "COMMIT_HISTORY_FOR_LLM.md").write_text(snap)
+
+    cmd = [
+        "gemini",
+        "--all-files",
+        "--model", model_name,
+        "--output", str(analysis_outfile),
+    ]
+    if show_memory:
+        cmd.append("--show-memory-usage")
+
+    if verbose:
+        _print_header("Starting Gemini CLI")
+        _print_stat("CWD", str(filtered_workspace))
+        _print_stat("Model", model_name)
+        _print_stat("Output", str(analysis_outfile))
+
+    stop = threading.Event()
+    t = threading.Thread(target=_spinner, args=(stop, "Gemini analysis"), daemon=True)
+    t.start()
+    try:
+        cp = subprocess.run(cmd, cwd=str(filtered_workspace), capture_output=True, text=True)
+    finally:
+        stop.set()
+        t.join()
+
+    # Always write whatever stdout produced; helpful for debugging
+    try:
+        if cp.stdout:
+            Path(analysis_outfile).write_text(cp.stdout)
+    except Exception as e:
+        if verbose:
+            _print_warn(f"could not write analysis output: {e}")
+
+    if cp.returncode != 0:
+        _print_err(f"Gemini exited with {cp.returncode}")
+        if verbose and cp.stderr:
+            _print_err(cp.stderr)
+    else:
+        if verbose:
+            _print_ok("Gemini analysis complete")
+
+    return cp.returncode
+
+
+# =====================================================================================
+# Size tree computation & rendering
+# =====================================================================================
+
+def _human_bytes(n: int) -> str:
+    """Human-readable bytes, no colors; aligned later."""
+    units = ["B", "KiB", "MiB", "GiB", "TiB"]
+    x = float(n)
+    for u in units:
+        if x < 1024.0 or u == units[-1]:
+            if u == "B":
+                return f"{int(x)} {u}"
+            return f"{x:.1f} {u}"
+        x /= 1024.0
+    return f"{int(n)} B"
+
+def compute_filtered_dir_sizes(
+    source_dir: Path,
+    exclude_dirs: Set[str],
+    exclude_exts: Set[str],
+    exclude_files: Set[str],
+    remove_patterns: List[str],
+    keep_patterns: List[str],
+) -> Dict[Path, int]:
+    """
+    Return a map of directory -> cumulative size (bytes), honoring exclusions.
+    """
+    exclusion = Exclusion(set(exclude_dirs), set(exclude_exts), set(exclude_files), list(remove_patterns), list(keep_patterns))
+    sizes: Dict[Path, int] = {}
+
+    def should_descend(rel: str) -> bool:
+        return not (rel and exclusion.dir_is_excluded(rel))
+
+    def add_size(p: Path, sz: int) -> None:
+        # bubble up to root
+        while True:
+            sizes[p] = sizes.get(p, 0) + sz
+            if p == source_dir:
+                break
+            p = p.parent
+
+    for r, dirs, files in os.walk(source_dir):
+        root_path = Path(r)
+        rel_dir = _relpath(root_path, source_dir)
+        # prune dirs
+        for d in list(dirs):
+            rel_d = _relpath(root_path / d, source_dir)
+            if not should_descend(rel_d):
+                dirs.remove(d)
+        # files
+        for f in files:
+            p = root_path / f
+            rel = _relpath(p, source_dir)
+            if exclusion.file_is_excluded(rel):
+                continue
+            try:
+                sz = p.stat().st_size
+            except OSError:
+                continue
+            add_size(root_path, sz)
+
+    # Ensure every traversed (non-excluded) dir at least appears (size 0)
+    for r, dirs, _ in os.walk(source_dir):
+        root_path = Path(r)
+        rel_dir = _relpath(root_path, source_dir)
+        if should_descend(rel_dir):
+            sizes.setdefault(root_path, sizes.get(root_path, 0))
+
+    return sizes
+
+def _gather_tree_lines_by_depth(
+    root: Path,
+    dir_sizes: Dict[Path, int],
+    max_depth: int,
+) -> List[Tuple[int, Path]]:
+    """
+    Collect (depth, path) entries up to max_depth for dirs present in dir_sizes.
+    """
+    items: List[Tuple[int, Path]] = []
+    root_prefix = str(root)
+
+    for d in sorted(dir_sizes.keys(), key=lambda p: str(p)):
+        if not str(d).startswith(root_prefix):
+            continue
+        depth = len(d.relative_to(root).parts) if d != root else 0
+        if depth <= max_depth:
+            items.append((depth, d))
+    # Sort by (depth, name)
+    items.sort(key=lambda t: (t[0], str(t[1]).lower()))
+    return items
+
+def render_size_tree(
+    source_dir: Path,
+    dir_sizes: Dict[Path, int],
+    depth: Optional[int] = None,
+    auto_top_n: Optional[int] = None,
+) -> str:
+    """
+    Render a size tree. If auto_top_n is provided, we pick the top-N heaviest
+    directories (excluding the root), include their ancestors, then render
+    up to the deepest level among those selections. Otherwise, depth-limited
+    from the root (default depth=1).
+    """
+    if auto_top_n is not None and auto_top_n <= 0:
+        return ""
+
+    # Build a set of dirs to include
+    include_dirs: Set[Path] = set()
+
+    if auto_top_n:
+        # Sort children by size (exclude the root itself for ranking)
+        ranked = [(p, sz) for p, sz in dir_sizes.items() if p != source_dir]
+        ranked.sort(key=lambda t: t[1], reverse=True)
+        top = [p for p, _ in ranked[:auto_top_n]]
+
+        # include ancestors for visual continuity
+        for p in top:
+            while True:
+                include_dirs.add(p)
+                if p == source_dir:
+                    break
+                p = p.parent
+        # compute max depth we need to render
+        max_depth = 0
+        for p in include_dirs:
+            d = len(p.relative_to(source_dir).parts) if p != source_dir else 0
+            if d > max_depth:
+                max_depth = d
+        items = []
+        for p in sorted(include_dirs, key=lambda q: (len(q.relative_to(source_dir).parts) if q != source_dir else 0, str(q).lower())):
+            depth_here = len(p.relative_to(source_dir).parts) if p != source_dir else 0
+            items.append((depth_here, p))
+    else:
+        max_depth = 1 if depth is None else max(0, depth)
+        items = _gather_tree_lines_by_depth(source_dir, dir_sizes, max_depth)
+
+    # Size column width
+    size_strings = {p: _human_bytes(dir_sizes.get(p, 0)) for _, p in items}
+    width = max((len(s) for s in size_strings.values()), default=0)
+
+    # Render lines
+    lines: List[str] = []
+    title = "Size Tree (auto top-N)" if auto_top_n else f"Size Tree (depth={max_depth})"
+    lines.append("")
+    lines.append(cfmt("==", "magenta") + " " + cfmt(title, "bold"))
+
+    for depth_i, p in items:
+        name = p.name if p != source_dir else p.name or str(source_dir)
+        indent = "    " * depth_i
+        size_s = size_strings[p].rjust(width)
+        lines.append(f"{indent}{name}/  {size_s}")
+
+    return "\n".join(lines)
+
+
+# =====================================================================================
+# CLI
+# =====================================================================================
+
+def main(argv: Optional[List[str]] = None) -> int:
+    p = argparse.ArgumentParser(description="Package a repo for LLMs and optionally run Gemini analysis.")
+    p.add_argument("source", help="Source folder to package")
+    p.add_argument("-o", "--output", "-O", help="Output base path (no extension needed)", default=None)
+
+    p.add_argument("-f", "--file-mode", action="store_true", help="Write consolidated text file")
+    p.add_argument("-z", "--zip-mode", action="store_true", help="Write zip file")
+
+    p.add_argument("-x", "--exclude-dir", action="append", default=[], help="Exclude directory (repeatable)")
+    p.add_argument("-e", "--exclude-ext", action="append", default=[], help="Exclude extension with dot (repeatable)")
+    p.add_argument("-X", "--exclude-file", action="append", default=[], help="Exclude filename (repeatable)")
+    p.add_argument("-I", "--include-file", action="append", default=[], help="Force-include filename (repeatable)")
+    p.add_argument("-R", "--remove-pattern", action="append", default=[], help="Remove pattern (fnmatch) (repeatable)")
+    p.add_argument("-K", "--keep-pattern", action="append", default=[], help="Keep pattern (fnmatch) (repeatable)")
+
+    p.add_argument("-m", "--max-size", type=int, default=None, help="Max zip size in MiB (best-effort)")
+    p.add_argument("-p", "--preferences", default="", help="Comma list of preferred extensions to trim first")
+
+    p.add_argument("-F", "--flatten", action="store_true", help="Flatten files into a staging dir before packaging")
+    p.add_argument("-N", "--name-by-path", action="store_true", help="When flattening, rename files by path")
+    p.add_argument("-v", "--verbose", action="store_true", help="Verbose progress output")
+
+    p.add_argument("-P", "--preset", choices=sorted(PRESETS.keys()), help="Language preset (github-style)")
+
+    p.add_argument("-G", "--gemini", action="store_true", help="Run Gemini CLI analysis on filtered workspace")
+    p.add_argument("-M", "--model", default="gemini-2.5-flash", help="Gemini model")
+    p.add_argument("-C", "--include-commits", action="store_true", help="Include a compact commit snapshot")
+    p.add_argument("-L", "--commit-limit", type=int, default=20, help="Limit number of commits included")
+    p.add_argument("-S", "--show-memory", action="store_true", help="Show memory usage in Gemini CLI")
+
+    # New tree-size flags
+    p.add_argument(
+        "-T", "--size-tree",
+        nargs="?", const=1, type=int, default=None,
+        help="Print a directory size tree to depth (default 1 if value omitted)."
     )
-    parser.add_argument("-s", "--source", required=True, help="Folder to process")
-    parser.add_argument("-o", "--output", required=True, help="Base output path (e.g., 'myarchive' or 'myarchive.zip')")
-    
-    mode_group = parser.add_argument_group(title="Output Modes")
-    mode_group.add_argument(
-        "-F", "--file-mode", action="store_true",
-        help="Produce one consolidated text file."
-    )
-    mode_group.add_argument(
-        "-Z", "--zip-mode", action="store_true",
-        help="Produce a zip file. (Default if no mode is specified)"
+    p.add_argument(
+        "-A", "--auto-size-tree",
+        nargs="?", const=25, type=int, default=None,
+        help="Print top-N largest directories (default 25). If deeper, parents are included."
     )
 
-    filter_group = parser.add_argument_group(title="Filtering and Exclusion")
-    filter_group.add_argument(
-        "-xd", "--exclude-dir", nargs="*", default=list(DEFAULT_EXCLUDE_DIRS),
-        help=f"Directory names to exclude. Defaults: {', '.join(DEFAULT_EXCLUDE_DIRS)}"
-    )
-    filter_group.add_argument(
-        "-xe", "--exclude-ext", nargs="*", default=list(DEFAULT_EXCLUDE_EXTS),
-        help=f"File extensions to exclude. Defaults: {', '.join(DEFAULT_EXCLUDE_EXTS)}"
-    )
-    filter_group.add_argument(
-        "-xf", "--exclude-file", nargs="*", default=list(DEFAULT_EXCLUDE_FILES),
-        help=f"Specific filenames to exclude. Defaults: {', '.join(DEFAULT_EXCLUDE_FILES)}"
-    )
-    filter_group.add_argument(
-        "-if", "--include-file", nargs="*", default=[], 
-        help="Force-include specific files (overrides -xf if names match)."
-    )
-    filter_group.add_argument(
-        "-r", "--remove-patterns", nargs="*", default=[],
-        help="Glob patterns for filenames/dirnames to remove. Matched against names."
-    )
-    filter_group.add_argument(
-        "-k", "--keep-patterns", nargs="*", default=[],
-        help="Glob patterns for filenames/dirnames to keep (overrides remove-patterns)."
-    )
+    args = p.parse_args(argv)
 
-    zip_options_group = parser.add_argument_group(title="Zip Mode Options")
-    zip_options_group.add_argument(
-        "-ms", "--max-size", type=int,
-        help="Maximum zip file size in MB. If exceeded, files are pruned from source/flattened copy."
-    )
-    zip_options_group.add_argument(
-        "-p", "--preferences", nargs="*", default=[],
-        help="Preferred file extensions for deletion if zip exceeds max-size."
-    )
+    source_dir = Path(args.source).resolve()
+    if not source_dir.is_dir():
+        _print_err(f"Source is not a directory: {source_dir}")
+        return 2
 
-    format_group = parser.add_argument_group(title="Formatting Options")
-    format_group.add_argument(
-        "-f", "--flatten", action="store_true",
-        help="Flatten directory structure before packaging."
-    )
-    format_group.add_argument(
-        "-np", "--name-by-path", action="store_true",
-        help="When flattening, rename files to include their original relative path."
-    )
-    
-    parser.add_argument(
-        "-v", "--verbose", action="store_true", help="Enable verbose output."
-    )
-
-    args = parser.parse_args()
-
+    # Determine what to produce
     run_file_mode = args.file_mode
     run_zip_mode = args.zip_mode
     if not run_file_mode and not run_zip_mode:
-        run_zip_mode = True 
+        run_zip_mode = True  # default
 
-    final_exclude_files = set(args.exclude_file) - set(args.include_file)
-    final_exclude_dirs = set(args.exclude_dir)
-    final_exclude_exts = set(args.exclude_ext)
+    # Output naming
+    output_base = args.output or (source_dir.name)
+    output_path = Path(output_base)
+    output_dir = output_path.parent if output_path.parent != Path("") else Path.cwd()
+    output_dir.mkdir(parents=True, exist_ok=True)
 
-    output_path_arg = Path(args.output)
-    output_dir = output_path_arg.parent
-    output_dir.mkdir(parents=True, exist_ok=True) 
-    
-    if output_path_arg.name.startswith('.') and not output_path_arg.stem.startswith('.'):
-        base_name_for_output = output_path_arg.name
+    if output_path.name.startswith('.') and not output_path.stem.startswith('.'):
+        base_name = output_path.name
     else:
-        base_name_for_output = output_path_arg.stem if output_path_arg.stem else output_path_arg.name
+        base_name = output_path.stem if output_path.stem else output_path.name
 
-    actual_text_output_path: Path | None = None
-    actual_zip_output_path: Path | None = None
+    # Build exclusion sets
+    final_exclude_dirs = set(DEFAULT_EXCLUDE_DIRS)
+    final_exclude_exts = set(DEFAULT_EXCLUDE_EXTS)
+    final_exclude_files = set(DEFAULT_EXCLUDE_FILES)
+    final_remove_patterns: List[str] = []
+
+    if args.preset:
+        ps = PRESETS[args.preset]
+        final_exclude_dirs.update(ps.get("dirs", set()))
+        final_exclude_exts.update(ps.get("exts", set()))
+        final_exclude_files.update(ps.get("files", set()))
+        final_remove_patterns.extend(ps.get("patterns", []))
+
+    # Apply CLI excludes
+    final_exclude_dirs.update(args.exclude_dir or [])
+    final_exclude_exts.update(args.exclude_ext or [])
+    final_exclude_files.update(args.exclude_file or [])
+    final_remove_patterns.extend(args.remove_pattern or [])
+    final_keep_patterns = list(args.keep_pattern or [])
+
+    # Force-include files (remove from exclude set)
+    final_exclude_files -= set(args.include_file or [])
+
+    # Preferences list
+    preferences = [s.strip() for s in (args.preferences or "").split(",") if s.strip()]
+
+    # Create outputs
+    actual_text = None
+    actual_zip = None
 
     if run_file_mode and run_zip_mode:
-        original_suffix = output_path_arg.suffix.lower()
-        if original_suffix == ".txt":
-            actual_text_output_path = output_path_arg
-            actual_zip_output_path = output_dir / (base_name_for_output + ".zip")
-        elif original_suffix == ".zip":
-            actual_zip_output_path = output_path_arg
-            actual_text_output_path = output_dir / (base_name_for_output + ".txt")
+        # two outputs, infer extensions if not provided
+        suffix = output_path.suffix.lower()
+        if suffix == ".txt":
+            actual_text = output_dir / output_path.name
+            actual_zip = output_dir / (base_name + ".zip")
+        elif suffix == ".zip":
+            actual_zip = output_dir / output_path.name
+            actual_text = output_dir / (base_name + ".txt")
         else:
-            if original_suffix: 
-                rprint(f"[yellow]Warning:[/yellow] Output path '{args.output}' has an unrecognized extension ('{original_suffix}') for combined mode. "
-                      f"Using '{base_name_for_output}.txt' and '{base_name_for_output}.zip'.")
-            actual_text_output_path = output_dir / (base_name_for_output + ".txt")
-            actual_zip_output_path = output_dir / (base_name_for_output + ".zip")
+            actual_text = output_dir / (base_name + ".txt")
+            actual_zip = output_dir / (base_name + ".zip")
     elif run_file_mode:
-        if output_path_arg.suffix.lower() == ".txt":
-            actual_text_output_path = output_path_arg
-        else:
-            if output_path_arg.suffix:
-                 rprint(f"[yellow]Warning:[/yellow] Output path '{args.output}' has suffix '{output_path_arg.suffix}'. "
-                       f"File mode is active, final output will be '{base_name_for_output}.txt'.")
-            actual_text_output_path = output_dir / (base_name_for_output + ".txt")
-    elif run_zip_mode: 
-        if output_path_arg.suffix.lower() == ".zip":
-            actual_zip_output_path = output_path_arg
-        else:
-            if output_path_arg.suffix:
-                rprint(f"[yellow]Warning:[/yellow] Output path '{args.output}' has suffix '{output_path_arg.suffix}'. "
-                      f"Zip mode is active, final output will be '{base_name_for_output}.zip'.")
-            actual_zip_output_path = output_dir / (base_name_for_output + ".zip")
+        actual_text = output_dir / (base_name + ".txt")
+    elif run_zip_mode:
+        actual_zip = output_dir / (base_name + ".zip")
 
-    if actual_text_output_path:
-        rprint(f"[bold magenta]Starting text file mode.[/bold magenta] Output to: [cyan]{actual_text_output_path}[/cyan]")
+    if args.verbose:
+        _print_header("Packaging parameters")
+        _print_stat("Preset", args.preset or "(none)")
+        _print_stat("Exclude dirs", str(len(final_exclude_dirs)))
+        _print_stat("Exclude exts", str(len(final_exclude_exts)))
+        _print_stat("Exclude files", str(len(final_exclude_files)))
+        if final_remove_patterns:
+            _print_stat("Remove patterns", ", ".join(final_remove_patterns))
+        if final_keep_patterns:
+            _print_stat("Keep patterns", ", ".join(final_keep_patterns))
+
+    if actual_text:
         text_file_mode(
-            args.source, str(actual_text_output_path),
+            str(source_dir), str(actual_text),
             final_exclude_dirs, final_exclude_exts, final_exclude_files,
-            args.remove_patterns, args.keep_patterns,
+            final_remove_patterns, final_keep_patterns,
+            args.flatten, args.name_by_path, args.verbose
+        )
+    if actual_zip:
+        zip_folder(
+            str(source_dir), str(actual_zip),
+            final_exclude_dirs, final_exclude_exts, final_exclude_files,
+            final_remove_patterns, final_keep_patterns,
+            args.max_size, preferences,
             args.flatten, args.name_by_path, args.verbose
         )
 
-    if actual_zip_output_path:
-        rprint(f"[bold magenta]Starting zip mode.[/bold magenta] Output to: [cyan]{actual_zip_output_path}[/cyan]")
-        zip_folder(
-            args.source, str(actual_zip_output_path),
+    # Gemini analysis (on a filtered workspace)
+    if args.gemini:
+        ws = None
+        try:
+            ws = prepare_analysis_workspace(
+                source_dir,
+                final_exclude_dirs, final_exclude_exts, final_exclude_files,
+                final_remove_patterns, final_keep_patterns,
+                verbose=args.verbose,
+            )
+            out_md = output_dir / (base_name + ".md")
+            rc = run_gemini_analysis(
+                filtered_workspace=ws,
+                model_name=args.model,
+                analysis_outfile=out_md,
+                include_commits=args.include_commits,
+                commit_limit=args.commit_limit,
+                show_memory=args.show_memory,
+                verbose=args.verbose,
+            )
+            if args.verbose:
+                _print_stat("Gemini return code", str(rc))
+        finally:
+            if ws:
+                shutil.rmtree(ws, ignore_errors=True)
+
+    # --------------------------
+    # Final: print size tree
+    # --------------------------
+    st_depth: Optional[int] = args.size_tree
+    st_topn: Optional[int] = args.auto_size_tree
+    if st_topn is not None:
+        # auto wins if both present
+        st_depth = None
+
+    if st_depth is not None or st_topn is not None:
+        sizes = compute_filtered_dir_sizes(
+            source_dir,
             final_exclude_dirs, final_exclude_exts, final_exclude_files,
-            args.remove_patterns, args.keep_patterns,
-            args.max_size, args.preferences,
-            args.flatten, args.name_by_path, args.verbose
+            final_remove_patterns, final_keep_patterns,
         )
+        report = render_size_tree(source_dir, sizes, depth=st_depth, auto_top_n=st_topn)
+        if report.strip():
+            print(report)
+
+    return 0
+
+
+if __name__ == "__main__":
+    raise SystemExit(main())
